requirements.txt
txt
flask
requests
python-pptx
pdf2image
openai
langchain